requirements/dev.txt
-r runtime.txt

# Local dev UX
django-browser-reload==1.18.0
django-debug-toolbar==4.4.6

# Test/data helpers
Faker==17.3.0
pip-audit==2.9.0

# Legacy browser/robot automation toolchain (kept for compatibility)
robotframework==7.3.2
robotframework-bsnlibrary==1.0.0
robotframework-databaselibrary==1.2.4
robotframework-faker==5.0.0
robotframework-pythonlibcore==4.0.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.0.0
robotframework-sshlibrary==3.8.0
selenium==4.7.2

# Explicit pins used by above toolchain
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
cryptography==46.0.7
h11==0.14.0
idna==3.4
outcome==1.2.0
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
scp==0.14.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wrapt==1.14.1
wsproto==1.2.0
